+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Notificaiton subscriptions - Shows GiB unit for network in modal
+
 ## [2.33.3] - 2023-02-27
 
 ### Changed